Split MCP batch query handler tests (#4108)
