+++ changelog.d/unreleased/4108.internal.md
@@ -1,0 +1,18 @@
+---
+category: internal
+issues:
+  - 4108
+affected:
+  - src/CodeIndex/Mcp/McpToolHandlers.BatchQuery.cs
+  - src/CodeIndex/Mcp/McpToolHandlers.cs
+  - tests/CodeIndex.Tests/McpServerBatchQueryTests.cs
+  - tests/CodeIndex.Tests/McpServerToolsCallTests.cs
+---
+
+## English
+
+- **Split the MCP `batch_query` handler and focused tests (#4108)** — `batch_query` now has a dedicated MCP handler file and representative command-handler tests live in a focused batch-query test file, reducing the largest MCP handler and tool-call test clusters without changing behavior.
+
+## 日本語
+
+- **MCP `batch_query` handler と focused test を分割しました (#4108)** — `batch_query` を専用の MCP handler ファイルへ移し、代表的な command-handler test を batch-query 専用テストファイルへ分けることで、挙動を変えずに最大級の MCP handler / tool-call test 集中を小さくしました。
